Move config helpers in own module
