+++ crates/nu-protocol/src/config/helper.rs
@@ -1,0 +1,20 @@
+use crate::{Record, ShellError, Span, Value};
+use std::collections::HashMap;
+
+pub(super) fn create_map(value: &Value) -> Result<HashMap<String, Value>, ShellError> {
+    Ok(value
+        .as_record()?
+        .iter()
+        .map(|(k, v)| (k.clone(), v.clone()))
+        .collect())
+}
+
+pub fn extract_value<'record>(
+    name: &str,
+    record: &'record Record,
+    span: Span,
+) -> Result<&'record Value, ShellError> {
+    record
+        .get(name)
+        .ok_or_else(|| ShellError::MissingConfigValue(name.to_string(), span))
+}
